Fix tests to parts
