--- testing/tests/DevExpress.exporter/jspdf.dataGrid.tests.js
@@ -1,2 +1,0 @@
-import './jspdfParts/jspdf.dataGrid.tests.js';
-import './jspdfParts/v3/jspdf_v3.dataGrid.tests.js';
