+++ testing/tests/DevExpress.exporter/jspdf.dataGrid_old.tests.js
@@ -1,0 +1,1 @@
+import './jspdfParts/jspdf.dataGrid.tests.js';
